# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "pos-proxy acts as the backend for POS module in the localnetwork\nit handles all the POS operations also handles the offline mode\nscenarios",
    "title": "pos-proxy api",
    "version": "1.0.0"
  },
  "host": "localhost",
  "basePath": "/api/",
  "paths": {
...
    "/api/opera/rooms/": {
      "get": {
        "description": "returns a list of reservations on Opera",
        "tags": [
          "opera"
        ],
        "summary": "List Opera Rooms",
        "operationId": "listRooms",
        "parameters": [
          {
            "name": "inquiry",
            "in": "query",
            "required": true
          },
          {
            "name": "terminal",
            "in": "query",
            "required": true
          },
          {
            "name": "store",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "postListItem",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/postListItem"
              }
            }
          }
        }
      }
    },
...
    "/api/pos/fixeddiscount/": {
      "get": {
        "description": "returns a list of fixed discounts for a store",
        "tags": [
          "discounts"
        ],
        "summary": "List Fixed Discounts",
        "operationId": "listFixedDiscoints",
        "parameters": [
          {
            "type": "integer",
            "description": "filter discounts by id",
            "name": "id",
            "schema": {
              "description": "filter discounts by id",
              "type": "integer"
            }
          },
          {
            "type": "integer",
            "description": "filter discounts by store",
            "name": "store",
            "schema": {
              "description": "filter discounts by store",
              "type": "integer"
            }
          },
          {
            "type": "integer",
            "description": "filter discounts by cashier id",
            "name": "poscashier_id",
            "schema": {
              "description": "filter discounts by cashier id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "fixedDiscount",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/fixedDiscount"
              }
            }
          }
        }
      },
      "post": {
        "description": "creates a new fixed discount",
        "tags": [
          "discounts"
        ],
        "summary": "Create Fixed Discount",
        "operationId": "createFixedDiscount",
        "parameters": [
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/fixedDiscount"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "fixedDiscount",
            "schema": {
              "$ref": "#/definitions/fixedDiscount"
            }
          }
        }
      }
    },
    "/api/pos/fixeddiscount/{id}/": {
      "get": {
        "description": "returns an object of a FixedDiscount based on ID",
        "tags": [
          "discounts"
        ],
        "summary": "Get Fixed Discount",
        "operationId": "getFixedDiscount",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "fixedDiscount",
            "schema": {
              "$ref": "#/definitions/fixedDiscount"
            }
          }
        }
      },
      "put": {
        "description": "updates a fixed discount based on ID",
        "tags": [
          "discounts"
        ],
        "summary": "Update Fixed Discount",
        "operationId": "updateFixedDiscount",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/fixedDiscount"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "fixedDiscount",
            "schema": {
              "$ref": "#/definitions/fixedDiscount"
            }
          }
        }
      },
      "delete": {
        "description": "deletes a fixed discount by id from mongodb\nthen proxy the request to the backend",
        "tags": [
          "discounts"
        ],
        "summary": "Delete Fixed Discount",
        "operationId": "deleteFixedDiscounts",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/api/pos/opera/{id}/": {
      "delete": {
        "description": "deletes opera configuration by ID",
        "tags": [
          "opera"
        ],
        "summary": "Delete Opera Config",
        "operationId": "deleteConfig",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/api/pos/posinvoices/": {
      "get": {
        "description": "retrieves list of settled or open invoices\nthe settled invoices are being proxied to backend.\nthe result is returned in a paginated style",
        "tags": [
          "invoices"
        ],
        "summary": "List Invoices (simplified)",
        "operationId": "listInvoicesPaginated",
        "parameters": [
          {
            "type": "boolean",
            "name": "is_settled",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "paginatedResponse",
            "schema": {
              "$ref": "#/definitions/paginatedResponse"
            }
          }
        }
      },
      "post": {
        "description": "creates a new invoice or update an old one",
        "tags": [
          "invoices"
        ],
        "summary": "Submit Invoice",
        "operationId": "submitInvoice",
        "parameters": [
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/invoicePOSTRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "invoice",
            "schema": {
              "$ref": "#/definitions/invoice"
            }
          }
        }
      }
    },
    "/api/pos/posinvoices/bulksubmit/": {
      "post": {
        "description": "loops over list of invoices, creates or updates\nthem, then release the terminal that is locked.",
        "tags": [
          "invoices"
        ],
        "summary": "Bulk Submit Invoices",
        "operationId": "bulkSubmitInvoices",
        "parameters": [
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/bulkSubmitRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "bulkSubmitResponse",
            "schema": {
              "$ref": "#/definitions/bulkSubmitResponse"
            }
          }
        }
      }
    },
    "/api/pos/posinvoices/changetable/": {
      "post": {
        "description": "moves the selected invoices from table to another table",
        "tags": [
          "invoices"
        ],
        "summary": "Change Table",
        "operationId": "changeTable",
        "parameters": [
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/changeTableRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "invoice",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/invoice"
              }
            }
          }
        }
      }
    },
...
    "/api/pos/posinvoices/folio/": {
      "post": {
        "description": "sends invoice lineitems to FDM and increase printing counter",
        "tags": [
          "invoices"
        ],
        "summary": "Folio",
        "operationId": "folioInvoice",
        "parameters": [
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/invoicePOSTRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "invoice",
            "schema": {
              "$ref": "#/definitions/invoice"
            }
          }
        }
      }
    },
    "/api/pos/posinvoices/houseuse/": {
      "post": {
        "description": "pay the invoice and settle it as house use",
        "tags": [
          "invoices"
        ],
        "summary": "Houseuse",
        "operationId": "houseuse",
        "parameters": [
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/invoicePOSTRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "invoice",
            "schema": {
              "$ref": "#/definitions/invoice"
            }
          }
        }
      }
    },
    "/api/pos/posinvoices/refund/": {
      "post": {
        "description": "handles the refund scenario and cancel postings.\nCreating a new invoice with the canceled postings.",
        "tags": [
          "invoices"
        ],
        "summary": "Refund Invoice",
        "operationId": "refundInvoice",
        "parameters": [
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/refundInvoiceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "refundInvoiceResponse",
            "schema": {
              "$ref": "#/definitions/refundInvoiceResponse"
            }
          }
        }
      }
    },
    "/api/pos/posinvoices/split/": {
      "post": {
        "description": "splits invoice to new invoices",
        "tags": [
          "invoices"
        ],
        "summary": "Split",
        "operationId": "split",
        "parameters": [
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/splitInvoiceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "invoice",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/invoice"
              }
            }
          }
        }
      }
    },
    "/api/pos/posinvoices/wasteandvoid/": {
      "post": {
        "description": "Wastes a lineitem",
        "tags": [
          "invoices"
        ],
        "summary": "Waste \u0026 Void",
        "operationId": "wasteAndVoid",
        "parameters": [
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/invoicePOSTRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "invoice",
            "schema": {
              "$ref": "#/definitions/invoice"
            }
          }
        }
      }
    },
    "/api/pos/posinvoices/{invoice_number}/": {
      "get": {
        "description": "fetches invoice from the database by invoice number",
        "tags": [
          "invoices"
        ],
        "summary": "Get Invoice",
        "operationId": "getInvoice",
        "parameters": [
          {
            "type": "string",
            "name": "invoice_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "invoice",
            "schema": {
              "$ref": "#/definitions/invoice"
            }
          }
        }
      }
    },
    "/api/pos/posinvoices/{invoice_number}/cancelpostings/": {
      "post": {
        "description": "cancels payments of a paid invocie based on postings frontend ids",
        "tags": [
          "invoices"
        ],
        "summary": "Cancel Postings",
        "operationId": "cancelPostings",
        "parameters": [
          {
            "type": "integer",
            "name": "invoice_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/cancelPostingsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "posting",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/posting"
              }
            }
          }
        }
      }
    },
    "/api/pos/posinvoices/{invoice_number}/createpostings/": {
      "post": {
        "description": "creates pospayments and pospostinginformations on invoice",
        "tags": [
          "invoices"
        ],
        "summary": "Create Postings (pay invoice)",
        "operationId": "createPostings",
        "parameters": [
          {
            "type": "integer",
            "name": "invoice_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/invoicePOSTRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "invoicePOSTRequest",
            "schema": {
              "$ref": "#/definitions/invoicePOSTRequest"
            }
          }
        }
      }
    },
    "/api/pos/posinvoices/{invoice_number}/getlatestchanges/": {
      "post": {
        "description": "gets the invoice and checks if it's locked or not",
        "tags": [
          "invoices"
        ],
        "summary": "Get Invoice Latest Changes",
        "operationId": "getInvoiceLatestChanges",
        "parameters": [
          {
            "type": "integer",
            "name": "invoice_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "invoice",
            "schema": {
              "$ref": "#/definitions/invoice"
            }
          }
        }
      }
    },
    "/api/pos/posinvoices/{invoice_number}/unlock/": {
      "get": {
        "description": "removes invoice_number key from redis and make\nthe invoice available to be picked up again by cashiers",
        "tags": [
          "invoices"
        ],
        "summary": "Unlock Invoice",
        "operationId": "unlockInvoice",
        "parameters": [
          {
            "type": "string",
            "name": "invoice_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/pos/posinvoices/{invoice_number}/void/": {
      "post": {
        "description": "closes invoice and sets the void reason",
        "tags": [
          "invoices"
        ],
        "summary": "Void Invoice",
        "operationId": "voidInvoice",
        "parameters": [
          {
            "type": "string",
            "name": "invoice_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/invoicePOSTRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "invoice",
            "schema": {
              "$ref": "#/definitions/invoice"
            }
          }
        }
      }
    },
...
    "/api/pos/store/{id}/": {
      "get": {
        "description": "returns a POS store data based on ID",
        "tags": [
          "stores"
        ],
        "summary": "Get Store",
        "operationId": "getStore",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "store",
            "schema": {
              "$ref": "#/definitions/store"
            }
          }
        }
      }
    },
    "/api/pos/storedetails/{id}/": {
      "get": {
        "description": "returns a POS stores details based on ID",
        "tags": [
          "stores"
        ],
        "summary": "Get Store Details",
        "operationId": "getStoreDetails",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "storeDetails",
            "schema": {
              "$ref": "#/definitions/storeDetails"
            }
          }
        }
      }
    },
...
    "/api/pos/tables/{id}/": {
      "get": {
        "description": "returns a details of a table by id",
        "tags": [
          "tables"
        ],
        "summary": "Get Table",
        "operationId": "getTable",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "table",
            "schema": {
              "$ref": "#/definitions/table"
            }
          }
        }
      },
      "put": {
        "description": "updates table data by id",
        "tags": [
          "tables"
        ],
        "summary": "Update Table",
        "operationId": "updateTable",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "type": "object",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/table"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "table",
            "schema": {
              "$ref": "#/definitions/table"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "fixedDiscount": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "number",
          "format": "double",
          "x-go-name": "Amount"
        },
        "cashiers": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "x-go-name": "CashierIDs"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "x-go-name": "ItemIDs"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "percentage": {
          "type": "number",
          "format": "double",
          "x-go-name": "Percentage"
        },
        "stores": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "x-go-name": "StoreIDs"
        }
      },
      "x-go-name": "FixedDiscount",
      "x-go-package": "pos-proxy/pos/models"
    },
...
    "postListItem": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "x-go-name": "FirstName"
        },
        "id": {
          "type": "string",
          "x-go-name": "ReservationId"
        },
        "last_name": {
          "type": "string",
          "x-go-name": "LastName"
        },
        "room_number": {
          "type": "string",
          "x-go-name": "RoomNumber"
        }
      },
      "x-go-name": "PostListItem",
      "x-go-package": "pos-proxy/opera"
    },
...
  }
}